Remove boilerplate fluff, fix logout & closeStyleMappings
Removed some boilerplate stuff as it will never be shown and is for quick start troubleshooting, logout and closeStyleMappings had broken as i was trying to de-populate old style selector names, fixed to de-populate all style dropdowns on style menu close. Also renamed the taskpane (i think). And did some stylizing of the UI for style mappings as well as the main page.