Add support for 1.1.1.1. format lists and fix whitespace issue with them
There is an odd whitespace issue with the 1.1.1. format lists because the margin between the text and the list symbol is different for each level and sometimes very large. Now it is filtered out, and looks flat, but maintains the number formatting for organization. Can support as far as word takes it.