Fix table parsing approach for test steps
Had an error where false positives for checking whether something was likely a table was occurring. Fixed my approach to use the API to determine whether a "line" is part of a table. There can still be overlap causing deleted lines from the description if the last line of a description is exactly equivalent to a line in the table, but Im unsure how to further fix that issue. General checking logic now contained in test() in taskpane.js.