Adds bolding for next step
This makes it so that the text describing the next step in the send to Spira sequence is bolded, including after errors occur. This also refactors the 'project-select' html element to 'product-select', fixes the spacing issue on the styles menu, and adds a small shadow to the pop-up, and makes the pop-up appearance more consistent across different screen sizes.