Add invalid image checking for test cases
Add checking for invalid images - images in tables which are not used within the imported data. These are "invalid" as they will not be used, but with the existing approach, they will mess up the order in which images are placed (the invalid image will be placed somewhere it shouldnt be, and offset all subsequent images to 1 place off of where they should be).