Fixes pluralization issue
Makes it so that the words "Requirement" and "Test Case" are plural when they should be for the confirmation and successfully sent screen. The in-progress sending screen  is still always plural, but that rarely appears for more than half a second when sending a single requirement, so I think it's acceptable.